+++ lib/multi_json/adapter_selector.rb
@@ -37,7 +37,7 @@ module MultiJson
     # @example
     #   AdapterSelector.default_adapter  #=> :oj
     def default_adapter
-      Concurrency::DEFAULT_ADAPTER.synchronize { @default_adapter ||= detect_best_adapter }
+      Concurrency.synchronize_default_adapter { @default_adapter ||= detect_best_adapter }
     end
 
     # Returns the default adapter class, excluding the given adapter name
@@ -52,7 +52,7 @@ module MultiJson
     # @example
     #   AdapterSelector.default_adapter_excluding(:fast_jsonparser)  #=> MultiJson::Adapters::Oj
     def default_adapter_excluding(excluded)
-      Concurrency::DEFAULT_ADAPTER.synchronize do
+      Concurrency.synchronize_default_adapter do
         name = loaded_adapter(excluding: excluded)
         name ||= installable_adapter(excluding: excluded)
         name ||= fallback_adapter