+++ lib/multi_json/options.rb
@@ -57,7 +57,7 @@ module MultiJson
     # @api private
     # @return [Hash] frozen empty hash
     def default_load_options
-      Concurrency::DEFAULT_OPTIONS.synchronize { @default_load_options ||= EMPTY_OPTIONS }
+      Concurrency.synchronize_default_options { @default_load_options ||= EMPTY_OPTIONS }
     end
 
     # Get default dump options
@@ -65,7 +65,7 @@ module MultiJson
     # @api private
     # @return [Hash] frozen empty hash
     def default_dump_options
-      Concurrency::DEFAULT_OPTIONS.synchronize { @default_dump_options ||= EMPTY_OPTIONS }
+      Concurrency.synchronize_default_options { @default_dump_options ||= EMPTY_OPTIONS }
     end
 
     private